fix: address PR #248 review comments (webhooks, CRDs, sync, dockerfile)
- Inbound/Outbound webhook: also reject when neither count nor addresses
  is set, matching the CRD's 'exactly one of' rule (with unit tests).
- TrafficMirror: make spec.direction optional and default it to 'both' to
  match the underlying NodeNetworkConfig MirrorACL behavior.
- Collector: store GRE tunnel key as int64 with min=0/max=4294967295 so
  the full unsigned 32-bit GRE key range round-trips through the API.
- platform-metallb Dockerfile: copy the full cmd/platform-metallb/ dir
  and build via package path so additional .go files aren't silently
  excluded from the image.
- network-sync: make remote-client QPS, Burst and per-request Timeout
  configurable via flags (defaults preserve current behavior plus a 30s
  request timeout to prevent reconcile hangs against unreachable
  workload clusters).
- ClusterController: evict cached remote client when the kubeconfig
  Secret disappears, lacks the 'value' key, is empty, or fails to parse,
  so subsequent reconciles cannot operate on stale credentials.
- ClusterController: ignore Secrets named exactly '-kubeconfig' in the
  Secret-to-Cluster mapper to avoid pointless empty-name reconciles.

+++ api/v1alpha1/network-connector/collector_types.go
@@ -32,9 +32,12 @@ type CollectorSpec struct {
 	// +kubebuilder:validation:Enum=l3gre;l2gre
 	Protocol string `json:"protocol"`
 
-	// Key is the GRE tunnel key.
+	// Key is the GRE tunnel key. The full unsigned 32-bit range is permitted;
+	// stored as int64 so values above the signed 32-bit range round-trip cleanly.
 	// +optional
-	Key *uint32 `json:"key,omitempty"`
+	// +kubebuilder:validation:Minimum=0
+	// +kubebuilder:validation:Maximum=4294967295
+	Key *int64 `json:"key,omitempty"`
 
 	// MirrorVRF references a VRF for the mirror VRF.
 	// +kubebuilder:validation:Required
